fix: amend DB for new DB creation script
This patch migrates to mysql so that the new blank db creation script works. Will break ARM based devices temporarily.
BREAKING: this will break arm based systems as this changes the db back to mysql. This is unavoidable as the new blank db script has variables types that do not seem to work with mariadb
will have to sort a solution for arm based devices shortly.