feat: create new db setup script on every fresh run
This patch amends how the container starts a new db on startup.
Previously it would use the current setup.sql in the aspen-dev-box repo
which required periodic running og the newSQL creation script.
This patch instead creates the file from the current blank db script in
the aspen repo (which should be regenerated with each release). It then
runs a seperate script with the account-profile and indexing profile
additions needed for adb. Any pending db updates are still then run in
the startup script.